グラフのリファクタリング
